refactor:(gateway) Make missed code residing directory.
