+++ mp-gateway/src/main/java/ru/pavbatol/myplace/gateway/security/auth/controller/AdminAuthController.java
@@ -1,4 +1,4 @@
-package ru.pavbatol.myplace.security.auth.controller;
+package ru.pavbatol.myplace.gateway.security.auth.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
@@ -7,8 +7,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import ru.pavbatol.myplace.app.api.ApiResponse;
-import ru.pavbatol.myplace.security.client.SecurityClient;
+import ru.pavbatol.myplace.gateway.app.api.ApiResponse;
+import ru.pavbatol.myplace.gateway.security.client.SecurityClient;
 
 import java.util.UUID;
