+++ mp-gateway/src/main/java/ru/pavbatol/myplace/gateway/security/auth/service/AuthService.java
@@ -1,0 +1,4 @@
+package ru.pavbatol.myplace.gateway.security.auth.service;
+
+public interface AuthService {
+}
