--- mp-gateway/src/main/java/ru/pavbatol/myplace/security/auth/service/AuthServiceImpl.java
@@ -1,4 +1,0 @@
-package ru.pavbatol.myplace.security.auth.service;
-
-public class AuthServiceImpl implements AuthService {
-}
